pyproject.toml
[tool.poetry]
name = "vorkers"
version = "0.1.0"
description = ""
authors = ["Никита Шидогубов <nikita@mer1d1an.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
gunicorn = "^23.0.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
